stopped the barrel at the wall.
next step is to implement the ladders and the ladder mechanic.

going to have to reconfigure collision but it'll be fun.

idea is that there will be red and green ladders, only mario can use red and only luigi can use green.

going to draw in donkey kong and the win platform after gravity has been implemented along with the ladder barrels

after all of that is done, i can start on points and lives and revival

so honestly so far so good!!!!